# gateway/docs/swagger.yaml
basePath: /
definitions:
  handlers.BulkAddTransactionsResponse:
    properties:
      failed:
        type: integer
      success:
        type: integer
    type: object
  handlers.authRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.authResponse:
    properties:
      token:
        type: string
    type: object
  internal.BudgetResponse:
    properties:
      category:
        type: string
      limit:
        type: number
      period:
        type: string
    type: object
  internal.CreateBudgetRequest:
    properties:
      category:
        type: string
      limit:
        type: number
      period:
        type: string
    type: object
  internal.CreateTransactionRequest:
    properties:
      amount:
        type: number
      category:
        type: string
      date:
        description: YYYY-MM-DD
        type: string
      description:
        type: string
    type: object
  internal.TransactionResponse:
    properties:
      amount:
        type: number
      category:
        type: string
      date:
        type: string
      description:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: GoFinance
  description: HTTP Gateway for personal finance microservices
  title: GoFinance Gateway API
  version: "1.0"
paths:
  /api/budgets:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal.BudgetResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List budgets
      tags:
      - budgets
    post:
      consumes:
      - application/json
      parameters:
      - description: Budget
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.CreateBudgetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Create budget
      tags:
      - budgets
  /api/reports/summary:
    get:
      parameters:
      - description: From date (YYYY-MM-DD)
        in: query
        name: from
        required: true
        type: string
      - description: To date (YYYY-MM-DD)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: float64
              type: number
            type: object
      security:
      - BearerAuth: []
      summary: Expense summary
      tags:
      - reports
  /api/transactions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal.TransactionResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      parameters:
      - description: Transaction
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - transactions
  /api/transactions/bulk:
    post:
      consumes:
      - application/json
      parameters:
      - description: Transactions
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/internal.CreateTransactionRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BulkAddTransactionsResponse'
      security:
      - BearerAuth: []
      summary: Bulk create transactions
      tags:
      - transactions
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user and get JWT
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.authRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.authResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.authRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.authResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"